pyproject.toml
[tool.poetry]
name = "python-scripts"
version = "0.1.0"
description = ""
authors = ["Your Name <your@email.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
playwright = "*"
pytest-playwright = "*"
allure-pytest = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-v --alluredir=allure-results"
pythonpath = ["."]
